refactor(rust): introduce `ModuleFactory` trait
