+++ crates/rolldown/src/ecmascript/mod.rs
@@ -1,2 +1,3 @@
 pub mod ecma_generator;
+pub mod ecma_module_factory;
 mod format;
