+++ crates/rolldown/src/types/mod.rs
@@ -7,6 +7,7 @@ pub mod bundle_output;
 pub mod bundler_fs;
 pub mod generator;
 pub mod linking_metadata;
+pub mod module_factory;
 pub mod namespace_alias;
 pub mod oxc_parse_type;
 pub mod symbols;
